feat: Add EventStoreDB adapter
* Added [EventStore DB](https://www.eventstore.com/eventstoredb)
  adapter
* Implement metadata propagation (MessageId/EventId, CorrelationId,
  CausationId) between API, EventStore, and Message Bus
* Added Docker Compose configuration to scale out application instance
  (3 replicas)
