+++ docker-compose.app-scaled.yaml
@@ -1,0 +1,92 @@
+version: "3.8"
+
+services:
+
+  app1:
+    depends_on:
+      - rabbitmq
+    build:
+      context: ./src/EventSourcedPM
+      dockerfile: eventsourced-pm.Dockerfile
+    environment:
+      - EventSourcedPM__ApiBaseUrl=http://*:43210
+      - EventSourcedPM__EventStore__Endpoint__Host=eventstore
+      - EventSourcedPM__EventStore__Username=admin
+      - EventSourcedPM__EventStore__Password=changeit
+      - EventSourcedPM__Postgres__Endpoint__Host=postgres
+      - EventSourcedPM__Postgres__Username=admin
+      - EventSourcedPM__Postgres__Password=changeit
+      - EventSourcedPM__Postgres__Database=espm
+      - EventSourcedPM__RabbitMq__Endpoint__Host=rabbitmq
+      - EventSourcedPM__RabbitMq__Username=admin
+      - EventSourcedPM__RabbitMq__Password=changeit
+      - EventSourcedPM__RabbitMq__VHost=espm
+      - EventSourcedPM__WaitForInfrastructureOnStartup=true
+    ports:
+      - 43210:43210
+    networks:
+      - devnet
+
+  app2:
+    depends_on:
+      - rabbitmq
+    build:
+      context: ./src/EventSourcedPM
+      dockerfile: eventsourced-pm.Dockerfile
+    environment:
+      - EventSourcedPM__ApiBaseUrl=http://*:43211
+      - EventSourcedPM__EventStore__Endpoint__Host=eventstore
+      - EventSourcedPM__EventStore__Username=admin
+      - EventSourcedPM__EventStore__Password=changeit
+      - EventSourcedPM__Postgres__Endpoint__Host=postgres
+      - EventSourcedPM__Postgres__Username=admin
+      - EventSourcedPM__Postgres__Password=changeit
+      - EventSourcedPM__Postgres__Database=espm
+      - EventSourcedPM__RabbitMq__Endpoint__Host=rabbitmq
+      - EventSourcedPM__RabbitMq__Username=admin
+      - EventSourcedPM__RabbitMq__Password=changeit
+      - EventSourcedPM__RabbitMq__VHost=espm
+      - EventSourcedPM__WaitForInfrastructureOnStartup=true
+    ports:
+      - 43211:43211
+    networks:
+      - devnet
+
+  app3:
+    depends_on:
+      - rabbitmq
+    build:
+      context: ./src/EventSourcedPM
+      dockerfile: eventsourced-pm.Dockerfile
+    environment:
+      - EventSourcedPM__ApiBaseUrl=http://*:43212
+      - EventSourcedPM__EventStore__Endpoint__Host=eventstore
+      - EventSourcedPM__EventStore__Username=admin
+      - EventSourcedPM__EventStore__Password=changeit
+      - EventSourcedPM__Postgres__Endpoint__Host=postgres
+      - EventSourcedPM__Postgres__Username=admin
+      - EventSourcedPM__Postgres__Password=changeit
+      - EventSourcedPM__Postgres__Database=espm
+      - EventSourcedPM__RabbitMq__Endpoint__Host=rabbitmq
+      - EventSourcedPM__RabbitMq__Username=admin
+      - EventSourcedPM__RabbitMq__Password=changeit
+      - EventSourcedPM__RabbitMq__VHost=espm
+      - EventSourcedPM__WaitForInfrastructureOnStartup=true
+    ports:
+      - 43212:43212
+    networks:
+      - devnet
+
+### NGINX load balancer
+#  proxy:
+#    depends_on:
+#      - app1
+#      - app2
+#      - app3
+#    build:
+#      context: ./proxy
+#      dockerfile: proxy.Dockerfile
+#    ports:
+#      - 43210:43210
+#    networks:
+#      - devnet
