+++ src/EventSourcedPM.Application/Orchestration/ShipmentProcessTriggersConsumer.cs
@@ -48,105 +48,240 @@ public class ShipmentProcessTriggersConsumer(IShipmentProcessManager processMana
 // and to allow changing process definition without having to (re-)wire triggers
 {
     public Task Consume(ConsumeContext<ProcessShipment> context) =>
-        processManager.InitializeProcess(context.Message);
+        processManager.InitializeProcess(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentProcessStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ManifestationAndDocumentsStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ManifestationAndDocumentsCompleted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ManifestationAndDocumentsFailed> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CustomsInvoiceGenerationStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(
         ConsumeContext<ManifestationAndDocumentsEvents.CustomsInvoiceGenerated> context
-    ) => processManager.InvokeProcessTrigger(context.Message);
+    ) =>
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CustomsInvoiceGenerationCompleted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CustomsInvoiceGenerationFailed> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentManifestationStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(
         ConsumeContext<ManifestationAndDocumentsEvents.ShipmentManifested> context
-    ) => processManager.InvokeProcessTrigger(context.Message);
+    ) =>
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(
         ConsumeContext<ManifestationAndDocumentsEvents.ShipmentManifestationFailed> context
-    ) => processManager.InvokeProcessTrigger(context.Message);
+    ) =>
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentManifestationCompleted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentManifestationFailed> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentLabelsGenerationStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(
         ConsumeContext<ManifestationAndDocumentsEvents.ShipmentLabelsGenerated> context
-    ) => processManager.InvokeProcessTrigger(context.Message);
+    ) =>
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentLabelsGenerationCompleted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentLabelsGenerationFailed> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ReceiptGenerationStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(
         ConsumeContext<ManifestationAndDocumentsEvents.ShipmentReceiptGenerated> context
-    ) => processManager.InvokeProcessTrigger(context.Message);
+    ) =>
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ReceiptGenerationCompleted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ReceiptGenerationFailed> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CombinedDocumentGenerationStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(
         ConsumeContext<ManifestationAndDocumentsEvents.ShipmentCombinedDocumentGenerated> context
-    ) => processManager.InvokeProcessTrigger(context.Message);
+    ) =>
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CombinedDocumentGenerationCompleted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CombinedDocumentGenerationFailed> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CollectionBookingStarted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CollectionBookingEvents.CollectionBooked> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(
         ConsumeContext<CollectionBookingEvents.CollectionBookingSubprocessFailed> context
-    ) => processManager.InvokeProcessTrigger(context.Message);
+    ) =>
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CollectionBookingCompleted> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<CollectionBookingFailed> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 
     public Task Consume(ConsumeContext<ShipmentProcessCompletionChecked> context) =>
-        processManager.InvokeProcessTrigger(context.Message);
+        processManager.InvokeProcessTrigger(
+            context.Message,
+            context.CorrelationId,
+            context.MessageId
+        );
 }
